refactoring(step1) : 피드백 사항 반영
