+++ src/main/java/domain/LottoSeller.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static domain.Lotto.getCountOfBuy;
+import static common.Constant.LOTTO_PRICE;
 
 public class LottoSeller {
 
@@ -13,8 +13,12 @@ public class LottoSeller {
     public List<Lotto> buyRandomNumberLottos(int money) {
         checkArgument(money);
         return IntStream.range(0, getCountOfBuy(money))
-                        .mapToObj(i -> new Lotto(generator.createRandomLottoNumbers()))
+                        .mapToObj(i -> generator.createRandomLottoNumbers())
                         .collect(Collectors.toList());
+    }
+
+    private int getCountOfBuy(int money) {
+        return money / LOTTO_PRICE;
     }
 
     private void checkArgument(int money) {
